fix: 찜은 본인 것만 접근 가능함

+++ dibs/views.py
@@ -1,8 +1,10 @@
 from django.db import IntegrityError
 from django.db.models import QuerySet
 from django.shortcuts import get_object_or_404
+from jwt import exceptions
 from rest_framework import viewsets, status
 from rest_framework.decorators import action
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 
 from dibs.models import DibsGroup, DibsDetail
@@ -14,6 +16,7 @@ from user.models import User
 # Create your views here.
 class DibsGroupViewSet(viewsets.GenericViewSet):
     serializer_class = DibsGroupSerializer
+    permission_classes = [IsAuthenticated]
 
     def get_queryset(self):
         queryset = DibsGroup.objects.prefetch_related('dibsDetails__product').all()
@@ -21,8 +24,10 @@ class DibsGroupViewSet(viewsets.GenericViewSet):
 
     @staticmethod
     def create(request, *args, **kwargs):
-        userId = request.data["userId"]
-        user = get_object_or_404(User, pk=userId)
+        if request.user.id:
+            user = request.user  # 현재 로그인한 사용자 정보를 사용
+        else:
+            raise exceptions.InvalidTokenError
 
         dibsGroup = DibsGroup.objects.create(
             name=request.data['name'],
@@ -33,7 +38,16 @@ class DibsGroupViewSet(viewsets.GenericViewSet):
 
     @staticmethod
     def destroy(request, pk: int) -> Response:
+        if request.user.id:
+            user = request.user  # 현재 로그인한 사용자 정보를 사용
+        else:
+            return Response("unauthorized", status=status.HTTP_401_UNAUTHORIZED)
+
         dibsGroup = get_object_or_404(DibsGroup, id=pk)
+
+        if user.id != dibsGroup.user.id:
+            return Response("unauthorized", status=status.HTTP_401_UNAUTHORIZED)
+
         dibsGroup.delete()
 
         return Response(status=status.HTTP_200_OK)
@@ -41,8 +55,9 @@ class DibsGroupViewSet(viewsets.GenericViewSet):
     # @action(["GET"], False, url_path=r"list")
     def list(self, request) -> Response:
         querySet: QuerySet[DibsGroup] = (
-            self.get_queryset()
+            self.get_queryset().filter(user__id=request.user.id)
         )
+
         serializer = self.get_serializer(querySet, many=True)
         return Response(serializer.data)
 
@@ -56,6 +71,14 @@ class DibsGroupViewSet(viewsets.GenericViewSet):
             # prefetch_related & select_related 쓰면 n+1 을 막을 수 있다~!
         )
 
+        if request.user.id:
+            user = request.user  # 현재 로그인한 사용자 정보를 사용
+        else:
+            raise exceptions.InvalidTokenError
+
+        if user.id != dibsGroup.user.id:
+            return Response("unauthorized", status=status.HTTP_401_UNAUTHORIZED)
+
         serializer = self.get_serializer(dibsGroup)
         return Response(serializer.data)
 
@@ -68,8 +91,16 @@ class DibsDetailViewSet(viewsets.GenericViewSet):
 
     @staticmethod
     def create(request, *args, **kwargs):
+        if request.user.id:
+            user = request.user  # 현재 로그인한 사용자 정보를 사용
+        else:
+            raise exceptions.InvalidTokenError
+
         dibsGroupId = request.data["dibsGroupId"]
         dibsGroup = get_object_or_404(DibsGroup, pk=dibsGroupId)
+
+        if user.id != dibsGroup.user.id:
+            return Response("unauthorized", status=status.HTTP_401_UNAUTHORIZED)
 
         productId = request.data["productId"]
         product = get_object_or_404(Product, pk=productId)
@@ -86,7 +117,16 @@ class DibsDetailViewSet(viewsets.GenericViewSet):
 
     @staticmethod
     def destroy(request, pk: int) -> Response:
+        if request.user.id:
+            user = request.user  # 현재 로그인한 사용자 정보를 사용
+        else:
+            raise exceptions.InvalidTokenError
+
         dibsDetail = get_object_or_404(DibsDetail, id=pk)
+
+        if user.id != dibsDetail.dibsGroup.user.id:
+            return Response("unauthorized", status=status.HTTP_401_UNAUTHORIZED)
+
         dibsDetail.delete()
 
         return Response(status=status.HTTP_200_OK)
